refactor: modify naming styled to styles and apply to Viewer, Drawer
