+++ libs/insight-viewer/src/Viewer/Viewer.styles.ts
